TangentImageHandler Class

TangentImageHandler Class

TangentImageHandler Class

Details

This class handles tangent images on a manifold. It provides methods to set a reference point, compute tangents, and perform various operations using a provided metric. # nolint: line_length_linter. Initialize the TangentImageHandler

Error if the tangent images have not been specified

Error if the reference point is not an object of class dppMatrix # nolint: line_length_linter

Error if the matrix is not of type dspMatrix Tangent images getter

Active bindings

ref_point

A matrix of type dppMatrix

tangent_images

A list of dspMatrix objects

Methods

Public methods

TangentImageHandler$new()

TangentImageHandler$set_reference_point()

TangentImageHandler$compute_tangents()

TangentImageHandler$compute_vecs()

TangentImageHandler$compute_conns()

TangentImageHandler$set_tangent_images()

TangentImageHandler$add_tangent_image()

TangentImageHandler$get_tangent_images()

TangentImageHandler$relocate_tangents()

TangentImageHandler$clone()

Method new()

Usage

TangentImageHandler$new(metric_obj, reference_point = NULL)

Arguments

metric_obj

An rmetric object for operations.

reference_point

An optional reference point on the manifold.

Returns

A new instance of TangentImageHandler. Set a new reference point.

Method set_reference_point()

If tangent images have been created, it recomputes them by mapping to the manifold and then to the new tangent space # nolint: line_length_linter

Usage

TangentImageHandler$set_reference_point(new_ref_pt)

Arguments

new_ref_pt

A new reference point of class dppMatrix.

Returns

None. Computes the tangent images from the points in the manifold

Method compute_tangents()

Usage

TangentImageHandler$compute_tangents(manifold_points)

Arguments

manifold_points

A list of connectomes

Returns

None Computes vectorizations from tangent images

Method compute_vecs()

Usage

TangentImageHandler$compute_vecs()

Returns

A matrix, each row of which is a vectorization Computes connectomes from tangent images

Method compute_conns()

Usage

TangentImageHandler$compute_conns()

Returns

A list of connectomes Setter for the tangent images

Method set_tangent_images()

Usage

TangentImageHandler$set_tangent_images(reference_point, tangent_images)

Arguments

reference_point

A connectome

tangent_images

A list of tangent images

Returns

None Appends a matrix to the list of tangent images

Method add_tangent_image()

Usage

TangentImageHandler$add_tangent_image(image)

Arguments

image

Matrix to be added

Method get_tangent_images()

Usage

TangentImageHandler$get_tangent_images()

Returns

list of tangent matrices Wrapper for set_reference_point

Method relocate_tangents()

Usage

TangentImageHandler$relocate_tangents(new_ref_pt)

Arguments

new_ref_pt

The new reference point

Returns

None

Method clone()

The objects of this class are cloneable with this method.

Usage

TangentImageHandler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.