Compute the Frechet Mean

This function computes the Frechet mean of a sample using an iterative algorithm.# nolint: line_length_linter.

Usage

compute_frechet_mean(sample, tol = 0.05, max_iter = 20, lr = 0.2)

Arguments

sample

An object of class CSample containing the sample data.

tol

A numeric value specifying the tolerance for convergence. Default is 0.05.# nolint: line_length_linter.

max_iter

An integer specifying the maximum number of iterations. Default is 20.# nolint: line_length_linter.

lr

A numeric value specifying the learning rate. Default is 0.2.

Value

The computed Frechet mean.

Details

The function iteratively updates the reference point of the sample until the change in the reference point is less than the specified tolerance or the maximum number of iterations is reached. If the tangent images are not already computed, they will be computed before starting the iterations.# nolint: line_length_linter.

Examples

if (FALSE) { # \dontrun{
sample <- conn_sample$new(
    conns = list_of_dppMatrix_objects,
    metric = some_metric
)
compute_frechet_mean(sample, tol = 0.01, max_iter = 50, lr = 0.1)
} # }