Relocate Tangent Representations to a New Reference Point

Changes the reference point for tangent space representations on a Riemannian manifold.# nolint: line_length_linter.

Usage

relocate(old_ref, new_ref, images, met)

Arguments

old_ref

A reference point on the manifold to be replaced. Must be an object of class dppMatrix from the Matrix package.# nolint: line_length_linter.

new_ref

The new reference point on the manifold. Must be an object of class dppMatrix from the Matrix package.# nolint: line_length_linter.

images

A list of tangent representations relative to the old reference point. Each element in the list must be an object of class dspMatrix.# nolint: line_length_linter.

met

A metric object of class rmetric, containing functions for Riemannian operations (logarithmic map, exponential map, vectorization, and inverse vectorization).# nolint: line_length_linter.

Value

A list of tangent representations relative to the new reference point. Each element in the returned list will be an object of class dspMatrix.# nolint: line_length_linter.